Changes to Tiled Rendering specifically for the Dragon Engine:
This system is very specifically tailored to this Engine and my own games. The only part of MonoGame.Extended that I'm interested in at this point is the Tiled integration/rendering, and there's a good chance that I'll strip this out of that repository and continue to maintain it as best as I can on my own as part of this Engine.

The main changes enabled a depth of something other than 0, which was the setting for all tiles in the Tilemap. When passing a position to the renderer, the previous Renderer only took a Vector2 and added a 0 as a Z value for passing into the Vertex buffer. I've modified it so that it takes a value between 0.25f and 0.75f along with it, such that each row of tiles renders higher than the previous one to emulate renderering bottom-up. This allows Tilemap sprites to slot in between the layers to have the correct placement within a scene.